+++ src/TRSConversion/Parse/ARI/Infeasibility.hs
@@ -3,12 +3,18 @@
 module TRSConversion.Parse.ARI.Infeasibility where
 
 import Control.Applicative (many)
-import TRSConversion.Parse.ARI.CTrs (parseAriCTrs', parseCondition)
-import TRSConversion.Parse.ARI.Utils (ARIParser, sExpr)
+import TRSConversion.Parse.ARI.CTrs (parseAriCTrs', parseCondition, pCondType)
+import TRSConversion.Parse.ARI.Utils (ARIParser, sExpr, keyword)
 import TRSConversion.Problem.CTrs.CTrs (CTrs (..), CondType, Condition, inferSigFromRules)
 import TRSConversion.Problem.CTrs.Infeasibility (Infeasibility (..))
 import TRSConversion.Problem.Trs.Sig (Sig)
 import qualified TRSConversion.Problem.Trs.TrsSig as Sig
+
+parseAriInfeasibility :: ARIParser (Infeasibility String String)
+parseAriInfeasibility = do
+  condType <- sExpr "format" $ keyword "CTRS" *> pCondType <* keyword ":problem" <* keyword "infeasibility"
+  parseAriInfeasibility' condType
+
 
 parseAriInfeasibility' :: CondType -> ARIParser (Infeasibility String String)
 parseAriInfeasibility' condType = do